requirements.txt
absl-py==1.0.0
anyio==3.6.1
aplus==0.11.0
argon2-cffi==20.1.0
asgiref==3.5.2
astropy==5.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
blake3==0.3.1
bleach==4.0.0
blis==0.7.5
bqplot==0.12.33
bubbly==1.0.2
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.1.0
cmdstanpy==0.9.5
colorama==0.4.4
commonmark==0.9.1
convertdate==2.3.2
cycler==0.10.0
cymem==2.0.6
Cython==0.29.26
dask==2022.7.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==4.0.6
django-windows-tools==0.2.1
entrypoints==0.3
ephem==4.1.3
fastai==2.5.3
fastapi==0.79.0
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
filelock==3.0.12
Flask==2.1.3
flatbuffers==2.0
fpdf2==2.5.5
frozendict==2.3.2
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h11==0.13.0
h2o==3.32.1.5
h5py==3.6.0
hijri-converter==2.2.2
holidays==0.12
httptools==0.4.0
hupper==1.10.3
idna==3.2
importlib-metadata==4.10.0
ipydatawidgets==4.3.1.post1
ipykernel==6.0.3
ipyleaflet==0.17.0
ipympl==0.9.1
ipython==7.26.0
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipyvue==1.7.0
ipyvuetify==1.8.2
ipywebrtc==0.6.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
joypy==0.2.5
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.2
jupyterlab-widgets==1.1.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
langcodes==3.3.0
libclang==12.0.0
llvmlite==0.38.1
locket==1.0.0
LunarCalendar==0.0.9
lxml==4.9.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.6
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numba==0.55.2
numpy==1.21.1
oauthlib==3.1.1
opt-einsum==3.3.0
outlier-utils==0.0.3
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
PasteDeploy==2.1.1
pathy==0.6.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2021.5.29
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==2.2.0
plotly==5.2.1
plotly-express==0.4.1
preshed==3.0.6
progressbar2==4.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.19.3
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyerfa==2.0.0.1
Pygments==2.9.0
PyMeeus==0.5.11
pyparsing==2.4.7
pyramid==2.0
pyrsistent==0.18.0
pystan==2.19.1.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
python-utils==3.3.3
pythreejs==2.3.0
pytz==2021.1
pywin32==301;platform_system == "Windows"
pywinpty==1.1.3;platform_system == "Windows"
PyYAML==6.0
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rich==12.5.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.7.1
setuptools-git==1.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
starlette==0.19.1
statsmodels==0.12.2
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.12.0
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
traittypes==0.2.1
translationstring==1.4
typer==0.4.0
typing-extensions==4.0.1
tzdata==2022.1
ujson==5.1.0
urllib3==1.26.6
uvicorn==0.18.2
vaex==4.9.2
vaex-astro==0.9.1
vaex-core==4.9.2
vaex-hdf5==0.12.2
vaex-jupyter==0.8.0
vaex-ml==0.17.0
vaex-server==0.8.1
vaex-viz==0.5.2
venusian==3.0.0
virtualenv==20.7.2
virtualenv-clone==0.5.6
wasabi==0.9.0
watchfiles==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.1.0
websockets==10.3
Werkzeug==2.0.2
widgetsnbextension==3.6.1
wrapt==1.13.3
xarray==2022.3.0
xgboost==1.4.2
XlsxWriter==3.0.3
xyzservices==2022.6.0
zipp==3.7.0
zope.deprecation==4.4.0
zope.interface==5.4.0
